Move shallow-equal test, fix eventemitter test
