+++ packages/next-server/lib/router/shallow-equals.js
