+++ test/unit/shallow-equal.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 
-import shallowEquals from 'next-server/dist/lib/shallow-equals'
+import shallowEquals from 'next-server/dist/lib/router/shallow-equals'
 
 describe('Shallow Equals', () => {
   it('should be true if both objects are the same', () => {
